+++ src/group-1.21.4-8/java/com/pvptweaks/config/PvpTweaksConfig.java
@@ -26,6 +26,7 @@ public class PvpTweaksConfig {
     private static PvpTweaksConfig LEGACY_INSTANCE = new PvpTweaksConfig();
 
     public static boolean useLegacyMenu = false;
+    public static boolean adjusterOpen = false;
     public int swordScalePct       = 70;
     public int axeScalePct         = 100;
     public int shieldScalePct      = 70;
@@ -36,6 +37,7 @@ public class PvpTweaksConfig {
     public int shieldRotY        = 0;
     public int shieldRotZ        = 0;
     public boolean shieldSampleShield = false;
+    public boolean shieldSampleActive = false;
     public SoundProfile soundShieldBreak = new SoundProfile();
     public int totemScalePct       = 100;
     public int goldenAppleScalePct = 100;
@@ -423,6 +425,7 @@ public class PvpTweaksConfig {
         cfg.shieldRotY = 0;
         cfg.shieldRotZ = 0;
         cfg.shieldSampleShield = false;
+        cfg.shieldSampleActive = false;
         cfg.totemScalePct = 100;
         cfg.goldenAppleScalePct = 100;
         cfg.anchorScalePct = 100;